pyproject.toml
[tool.poetry]
name = "emsug"
version = "0.0.0"
description = "a emergency management system using genai"
authors = ["Muhammad Fiaz <contact@muhammadfiaz.com>"]
repository = "https://github.com/muhammad-fiaz/Emergency-Management-System-using-Generative-AI.git"
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.8.1"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
langchain = "^0.3.2"
transformers = "^4.45.1"
bs4 = "^0.0.2"
logly = "^0.0.6"
sentence-transformers = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
